% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Create a Customized Bar Chart}
\usage{
bar_chart(data, x, y)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x}{A string representing the name of the x variable.}

\item{y}{A string representing the name of the y variable.}
}
\description{
This function creates a bar chart with ggplot2.
}
\examples{
bar_chart(mtcars, "cyl", "mpg")
}
